+++ CHANGES.rst
@@ -65,6 +65,7 @@ Improvements
 - Support cloning surveys when cloning events (:issue:`2045`, :pr:`4910`)
 - Show external contribution references in conferences (:issue:`4928`, :pr:`4933`)
 - Allow changing the rating scale in abstract/paper reviewing even after reviewing started (:pr:`4942`)
+- Allow blacklisting email addresses for user registrations (:issue:`4644`, :pr:`4946`)
 
 Bugfixes
 ^^^^^^^^